🐛FIX : Solve duplicate fetch comment edit API

+++ src/pages/Detail.tsx
@@ -67,13 +67,19 @@ export default function Detail() {
   const [displayScore, setDisplayScore] = useState<number | null>(score);
   const [bookmark, setBookmark] = useState(false);
   const [value, setValue] = useState('');
-  const [status, setStatus] = useState('readOnly');
+  //: 코멘트 상태
+  const [status, setStatus] = useState('null');
   const [comment, setComment] = useState<Comment | null>(null);
   const [comments, setComments] = useState<commentsList[]>([]);
   const [seeMore, setSeeMore] = useState(false);
   const [movieInfo, setMovieInfo] = useState<movieDetailInfoType | null>(null);
   const [isLike, setIsLike] = useState(false);
+
+  // const [isCommunicating, setIsCommunicating] = useState(false);
+  // const isCommunicating = useRef<boolean>(false);
+  let isCommunicating = false;
   const formtag = useRef(null);
+
   const editCommentInput = useRef<HTMLInputElement>(null);
 
   const navigate = useNavigate();
@@ -88,6 +94,12 @@ export default function Detail() {
   //   const token = useSelector((state) => state.auth.token);
   //   const account = useSelector((state) => state.auth.account);
   //   const userName = useSelector((state) => state.auth.userName);
+
+  /**
+   *
+   * 상태 : readonly, post, patch, delete, null
+   * api를 보내기전에 무슨상태인지 알아야한다. api 중복요청 안되게끔
+   */
 
   const requiredLogin = () => {
     alert('로그인 후 이용해 주세요');
@@ -139,6 +151,7 @@ export default function Detail() {
         );
         setComment(myComment.data.result);
         setValue(myComment.data.result.content);
+        setStatus('readOnly');
       } catch (error) {
         // console.log(error);
       }
@@ -186,6 +199,7 @@ export default function Detail() {
         const commentScore = await axios.get(
           `${LOCALAPI}/api/movie-reports/summary/movies/${movieId}`,
         );
+
         const addIsLike = commentScore.data.list.map((value: commentsList) => {
           return { ...value, isLike: false };
         });
@@ -317,10 +331,6 @@ export default function Detail() {
   //   console.log(response);
   // };
 
-  function addComment(e: any) {
-    setValue(e.target.value);
-  }
-
   async function like(
     e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
     commentId: number,
@@ -382,29 +392,71 @@ export default function Detail() {
     }
   }
 
+  function addComment(e: any) {
+    setValue(e.target.value);
+  }
+
+  //: input에서 enter눌렀을 때
+  function enterPressComment(
+    e: { key: string; preventDefault: () => void },
+    action: string,
+  ) {
+    if (e.key !== 'Enter') return;
+    if (status === 'readOnly') return;
+    e.preventDefault();
+    setStatus(action);
+  }
+
+  //: comment 등록, 수정, 취소, 삭제 버튼 눌렀을 때 action
+  function clickCommentButton(action: string) {
+    setStatus(action);
+  }
+
+  useEffect(() => {
+    if (status === 'create') {
+      handleAddComment();
+    } else if (status === 'not-readOnly') {
+      editComment();
+    } else if (status === 'edit') {
+      sendEditComment();
+    } else if (status === 'cancel') {
+      cancel();
+    } else if (status === 'delete') {
+      deleteComment();
+    }
+  }, [status]);
+
   async function handleAddComment() {
-    if (access_token) {
-      if (!value) return;
-      const removeSpace = value.trim();
-      try {
-        const response = await axios.post(
-          `${LOCALAPI}/api/movie-reports/${movieId}/comments`,
-          {
-            content: removeSpace,
-          },
-          bearer_header,
-        );
-        // setComments([...comments, response.data.data]);
-        const myComment = await axios.get(
-          `${LOCALAPI}/api/user-reports/movies/${movieId}/comment`,
-          bearer_header,
-        );
-        setComment(myComment.data.result);
-      } catch (error) {
-        console.log(error);
+    if (isCommunicating) return;
+
+    if (status === 'create') {
+      isCommunicating = true;
+      if (access_token) {
+        if (!value) return;
+        const removeSpace = value.trim();
+        try {
+          const response = await axios.post(
+            `${LOCALAPI}/api/movie-reports/${movieId}/comments`,
+            {
+              content: removeSpace,
+            },
+            bearer_header,
+          );
+          // setComments([...comments, response.data.data]);
+          const myComment = await axios.get(
+            `${LOCALAPI}/api/user-reports/movies/${movieId}/comment`,
+            bearer_header,
+          );
+          setComment(myComment.data.result);
+          setStatus('readOnly');
+        } catch (error) {
+          console.log(error);
+        } finally {
+          isCommunicating = false;
+        }
+      } else {
+        requiredLogin();
       }
-    } else {
-      requiredLogin();
     }
   }
 
@@ -412,12 +464,14 @@ export default function Detail() {
     if (editCommentInput.current) {
       editCommentInput.current.readOnly = false;
       editCommentInput.current.className = 'active';
-      setStatus('edit');
     }
   }
 
   async function sendEditComment() {
+    if (isCommunicating) return;
+
     if (status === 'edit') {
+      isCommunicating = true;
       try {
         const response = await axios.patch(
           `${LOCALAPI}/api/movie-reports/${movieId}/comments/${comment?.comment_id}`,
@@ -426,6 +480,7 @@ export default function Detail() {
         );
         if (response.status === 200) {
           setStatus('readOnly');
+          setComment(response.data.data);
           if (editCommentInput.current) {
             editCommentInput.current.className = '';
             editCommentInput.current.readOnly = true;
@@ -433,46 +488,42 @@ export default function Detail() {
         }
       } catch (error) {
         console.log(error);
+      } finally {
+        isCommunicating = false;
       }
     }
   }
 
   async function deleteComment() {
-    try {
-      const response = await axios.delete(
-        `${LOCALAPI}/api/movie-reports/${movieId}/comments/${comment?.comment_id}`,
-        bearer_header,
-      );
-      setComment(null);
-      setValue('');
-    } catch (error) {
-      console.log(error);
+    if (isCommunicating) return;
+
+    if (status === 'delete') {
+      isCommunicating = true;
+      try {
+        const response = await axios.delete(
+          `${LOCALAPI}/api/movie-reports/${movieId}/comments/${comment?.comment_id}`,
+          bearer_header,
+        );
+        setComment(null);
+        setValue('');
+        setStatus('null');
+      } catch (error) {
+        console.log(error);
+      } finally {
+        isCommunicating = false;
+      }
     }
   }
 
   function cancel() {
-    setStatus('readOnly');
     if (editCommentInput.current) {
       editCommentInput.current.className = '';
       editCommentInput.current.readOnly = true;
     }
-
     if (comment !== null) {
       setValue(comment.content);
     }
-  }
-
-  function enterPressComment(
-    e: { key: string; preventDefault: () => void },
-    act: string,
-  ) {
-    if (e.key !== 'Enter') return;
-    e.preventDefault();
-    if (act === 'create') {
-      handleAddComment();
-    } else {
-      sendEditComment();
-    }
+    setStatus('readOnly');
   }
 
   return (
@@ -562,9 +613,9 @@ export default function Detail() {
                   <div className="my-comment">
                     {comment ? (
                       <form
-                        action="/detail"
                         ref={formtag}
                         className="is-comment"
+                        onSubmit={(e) => e.preventDefault()}
                       >
                         <fieldset>
                           <legend>내가 작성한 코멘트</legend>
@@ -573,25 +624,41 @@ export default function Detail() {
                             value={value}
                             // placeholder={comment.content}
                             onChange={addComment}
-                            onKeyDown={(e) => enterPressComment(e, 'update')}
+                            onKeyDown={(e) => enterPressComment(e, 'edit')}
                             readOnly={true}
                             ref={editCommentInput}
                           />
                           {status === 'readOnly' ? (
+                            //: 읽기 전용일 때
                             <div>
-                              <button type="button" onClick={editComment}>
+                              <button
+                                type="button"
+                                onClick={() =>
+                                  clickCommentButton('not-readOnly')
+                                }
+                              >
                                 수정
                               </button>
-                              <button type="button" onClick={deleteComment}>
+                              <button
+                                type="button"
+                                onClick={() => clickCommentButton('delete')}
+                              >
                                 삭제
                               </button>
                             </div>
                           ) : (
+                            //: 수정 상태일 때
                             <div>
-                              <button type="button" onClick={cancel}>
+                              <button
+                                type="button"
+                                onClick={() => clickCommentButton('cancel')}
+                              >
                                 취소
                               </button>
-                              <button type="button" onClick={sendEditComment}>
+                              <button
+                                type="button"
+                                onClick={() => clickCommentButton('edit')}
+                              >
                                 완료
                               </button>
                             </div>
@@ -600,24 +667,23 @@ export default function Detail() {
                       </form>
                     ) : (
                       <form
-                        action="/detail"
                         ref={formtag}
                         className="no-comment"
+                        onSubmit={(e) => e.preventDefault()}
                       >
                         <fieldset>
                           <legend>코멘트 작성하기</legend>
                           <input
                             type="text"
                             value={value}
-                            placeholder="코멘트 작성 준비중입니다"
+                            placeholder="코멘트를 자유롭게 남겨주세요"
                             onChange={addComment}
                             onKeyDown={(e) => enterPressComment(e, 'create')}
                             // disabled
                           />
                           <button
                             type="button"
-                            onClick={handleAddComment}
-                            // disabled={true}
+                            onClick={() => clickCommentButton('create')} // disabled={true}
                             // className="unclickable"
                           >
                             코멘트 작성
